+++ tv-data/src/main/java/com/flimbis/tvmaze/data/datasource/remote/EpisodesRemoteDataSourceImpl.kt
@@ -8,7 +8,7 @@ import io.reactivex.Observable
  * Created by Fifi on 12/21/2017.
  */
 class EpisodesRemoteDataSourceImpl(val apiService: ApiService) : EpisodesDataSource {
-    override fun getAll(showId: Int): Observable<List<Episodes>> {
+    override fun getAll(showId: Long): Observable<List<Episodes>> {
         return apiService.getAllEpisodesOfShow(showId)
     }
 }
